LayoutTests/editing/selection/move-up-into-wrapped-line.html

<!DOCTYPE html>
<html>
<body>
<div style="width: 6ex;" contenteditable=true>
<div>hello world</div><div><br></div>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This test moves up caret into a wrapped line. \n'
 + 'i.e. the caret is moved from the empty line below "hello world" to before "world".');

var div = document.getElementsByTagName('div')[0];
div.focus();
getSelection().setPosition(div, div.childNodes.length);
getSelection().modify('move', 'backward', 'line');

Markup.dump(div);

</script>
</body>
</html>
