third_party/blink/web_tests/platform/generic/editing/selection/move-up-into-wrapped-line-expected.txt

This test moves up caret into a wrapped line. 
i.e. the caret is moved from the empty line below "hello world" to before "world".
| "
"
| <div>
|   "hello <#selection-caret>world"
| <div>
|   <br>
| "
"
